Current page files for my wix site
These are the 3 items that refer to my wix code database and pull in data from users using my form. 
It send them an email but it cuts off the email from sendGrid about 1/3 of the way through.

http://www.ca-rd.com/project-contact-form

the email result that clients get is attached and sadly cuts of. 

also I dont want to use sendgrid but its the only way I know on wix right now.